Fixed some issues

+++ SQL/sources/1_creation/table_poaInfo.sql
@@ -1,6 +1,6 @@
 CREATE TABLE Poa_Info (
     
-    poApplicationNum INTEGER(10) PRIMARY KEY,
+    poApplicationNum INTEGER PRIMARY KEY,
     poaTtiles VARCHAR (20),
     poaName VARCHAR (20)
